+++ pymongo/mongo_client.py
@@ -312,13 +312,15 @@ class MongoClient(common.BaseObject):
             https://docs.mongodb.com/manual/faq/diagnostics/#does-tcp-keepalive-time-affect-mongodb-deployments",
           - `compressors`: Comma separated list of compressors for wire
             protocol compression. The list is used to negotiate a compressor
-            with the server. Currently supported options are "snappy" and
-            "zlib". Support for snappy requires the
+            with the server. Currently supported options are "snappy", "zlib"
+            and "zstd". Support for snappy requires the
             `python-snappy <https://pypi.org/project/python-snappy/>`_ package.
-            zlib support requires the Python standard library zlib module.
-            By default no compression is used. Compression support must also be
-            enabled on the server. MongoDB 3.4+ supports snappy compression.
-            MongoDB 3.6+ supports snappy and zlib.
+            zlib support requires the Python standard library zlib module. zstd
+            requires the `zstandard <https://pypi.org/project/zstandard/>`_
+            package. By default no compression is used. Compression support
+            must also be enabled on the server. MongoDB 3.4+ supports snappy
+            compression. MongoDB 3.6 adds support for zlib. MongoDB 4.2 adds
+            support for zstd.
           - `zlibCompressionLevel`: (int) The zlib compression level to use
             when zlib is used as the wire protocol compressor. Supported values
             are -1 through 9. -1 tells the zlib library to use its default